Create Blanks.ino
first commit with a framework for a demo-mode Blank sketch. This suggests the demo modes will change based on a timer, but it might make sense for another variable to be responsible for switching mode (i.e. button press, the completion of a number of cycles or some other user interaction). This sketch should not be involved, but an easter egg here or there never hurt nobody 😛